Revert "Brush up code"
This reverts commit 12159c403955f54066ed8c532ed991f829edfc1f.
