+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/DataSource.scala
@@ -352,6 +352,9 @@ case class DataSource(
                 "It must be specified manually")
         }
 
+        SchemaUtils.checkSchemaColumnNameDuplication(
+          dataSchema, "in the datasource", sparkSession.sessionState.conf.caseSensitiveAnalysis)
+
         HadoopFsRelation(
           fileCatalog,
           partitionSchema = fileCatalog.partitionSchema,
