+++ sql/core/src/test/resources/sql-tests/results/create.sql.out
@@ -1,5 +1,5 @@
 -- Automatically generated by SQLQueryTestSuite
--- Number of queries: 5
+-- Number of queries: 9
 
 
 -- !query 0
@@ -20,26 +20,62 @@ Found duplicate column(s) in the table definition of `t`: `c1`, `c0`;
 
 
 -- !query 2
-SET spark.sql.caseSensitive=false
+CREATE TABLE t(c0 INT, c1 INT) PARTITIONED BY (c1 INT) STORED AS parquet
 -- !query 2 schema
-struct<key:string,value:string>
+struct<>
 -- !query 2 output
-spark.sql.caseSensitive	false
+org.apache.spark.sql.AnalysisException
+Found duplicate column(s) in the table definition of `t`: `c1`;
 
 
 -- !query 3
-CREATE TABLE t(c0 STRING, c1 INT, c1 DOUBLE, c0 INT) USING parquet
+CREATE TABLE t(c0 INT) PARTITIONED BY (c1 INT, c1 INT) STORED AS parquet
 -- !query 3 schema
 struct<>
 -- !query 3 output
 org.apache.spark.sql.AnalysisException
-Found duplicate column(s) in the table definition of `t`: `c1`, `c0`;
+Found duplicate column(s) in the table definition of `t`: `c1`;
 
 
 -- !query 4
-CREATE TABLE t(ab STRING, cd INT, ef DOUBLE, Ab INT) USING parquet
+SET spark.sql.caseSensitive=false
 -- !query 4 schema
-struct<>
+struct<key:string,value:string>
 -- !query 4 output
+spark.sql.caseSensitive	false
+
+
+-- !query 5
+CREATE TABLE t(c0 STRING, c1 INT, c1 DOUBLE, c0 INT) USING parquet
+-- !query 5 schema
+struct<>
+-- !query 5 output
+org.apache.spark.sql.AnalysisException
+Found duplicate column(s) in the table definition of `t`: `c1`, `c0`;
+
+
+-- !query 6
+CREATE TABLE t(ab STRING, cd INT, ef DOUBLE, Ab INT) USING parquet
+-- !query 6 schema
+struct<>
+-- !query 6 output
+org.apache.spark.sql.AnalysisException
+Found duplicate column(s) in the table definition of `t`: `ab`;
+
+
+-- !query 7
+CREATE TABLE t(ab INT, cd INT) PARTITIONED BY (Ab INT) STORED AS parquet
+-- !query 7 schema
+struct<>
+-- !query 7 output
+org.apache.spark.sql.AnalysisException
+Found duplicate column(s) in the table definition of `t`: `ab`;
+
+
+-- !query 8
+CREATE TABLE t(ab INT, cd INT) PARTITIONED BY (Ab INT, aB INT) STORED AS parquet
+-- !query 8 schema
+struct<>
+-- !query 8 output
 org.apache.spark.sql.AnalysisException
 Found duplicate column(s) in the table definition of `t`: `ab`;
